# [.NET] Avoid allocation and improve parsing time