--- a/dotnet/Gherkin.Specs/EventStubs/GherkinEventsProvider.cs
+++ b/dotnet/Gherkin.Specs/EventStubs/GherkinEventsProvider.cs
@@ -69,7 +69,7 @@ public class GherkinEventsProvider(bool printSource, bool printAst, bool printPi
                 Message = e.Message,
                 Source = new SourceReference()
                 {
-                    Location = new Location(e.Location.Column, e.Location.Line),
+                    Location = e.Location.HasValue ? new Location(e.Location.GetValueOrDefault().Column, e.Location.GetValueOrDefault().Line) : null,
                     Uri = uri
                 }
             }