--- a/dotnet/Gherkin.Specs/Tokens/TestTokenFormatter.cs
+++ b/dotnet/Gherkin.Specs/Tokens/TestTokenFormatter.cs
@@ -7,7 +7,8 @@ class TestTokenFormatter
         if (token.IsEOF)
             return "EOF";
 
-        string stepTypeText;
+        string stepTypeText = string.Empty;
+        string matchedItemsText = null;
         switch (token.MatchedType)
         {
             case TokenType.FeatureLine:
@@ -22,12 +23,13 @@ class TestTokenFormatter
                 var tokenType = token.MatchedGherkinDialect.GetStepKeywordType(token.MatchedKeyword);
                 stepTypeText = $"({tokenType})";
                 break;
-            default:
-                stepTypeText = "";
+            case TokenType.TagLine:
+                matchedItemsText = string.Join(",", token.Line.GetTags().Select(i => i.Column + ":" + i.Text));
+                break;
+            case TokenType.TableRow:
+                matchedItemsText = string.Join(",", token.Line.GetTableCells().Select(i => i.Column + ":" + i.Text));
                 break;
         }
-
-        var matchedItemsText = token.MatchedItems == null ? "" : string.Join(",", token.MatchedItems.Select(i => i.Column + ":" + i.Text));
 
         return $"({token.Location.Line}:{token.Location.Column}){token.MatchedType}:{stepTypeText}{token.MatchedKeyword}/{token.MatchedText}/{matchedItemsText}";
     }