--- a/dotnet/Gherkin/Ast/Background.cs
+++ b/dotnet/Gherkin/Ast/Background.cs
@@ -1,4 +1,4 @@
 namespace Gherkin.Ast;
 
-public class Background(Location location, string keyword, string name, string description, Step[] steps)
+public class Background(Location location, string keyword, string name, string description, IEnumerable<Step> steps)
     : StepsContainer(location, keyword, name, description, steps);