--- a/dotnet/Gherkin/Ast/GherkinDocument.cs
+++ b/dotnet/Gherkin/Ast/GherkinDocument.cs
@@ -1,6 +1,6 @@
 namespace Gherkin.Ast;
 
-public class GherkinDocument(Feature feature, Comment[] comments)
+public class GherkinDocument(Feature feature, IEnumerable<Comment> comments)
 {
     public Feature Feature { get; } = feature;
     public IEnumerable<Comment> Comments { get; } = comments;