--- a/dotnet/Gherkin/Ast/Location.cs
+++ b/dotnet/Gherkin/Ast/Location.cs
@@ -1,6 +1,6 @@
 namespace Gherkin.Ast;
 
-public class Location(int line = 0, int column = 0)
+public readonly struct Location(int line = 0, int column = 0)
 {
     public int Line { get; } = line;
     public int Column { get; } = column;