--- a/dotnet/Gherkin/Ast/Scenario.cs
+++ b/dotnet/Gherkin/Ast/Scenario.cs
@@ -1,6 +1,6 @@
 namespace Gherkin.Ast;
 
-public class Scenario(Tag[] tags, Location location, string keyword, string name, string description, Step[] steps, Examples[] examples)
+public class Scenario(IEnumerable<Tag> tags, Location location, string keyword, string name, string description, IEnumerable<Step> steps, IEnumerable<Examples> examples)
     : StepsContainer(location, keyword, name, description, steps), IHasTags
 {
     public IEnumerable<Tag> Tags { get; } = tags;