--- a/dotnet/Gherkin/Ast/TableCell.cs
+++ b/dotnet/Gherkin/Ast/TableCell.cs
@@ -1,6 +1,6 @@
 namespace Gherkin.Ast;
 
-public class TableCell(Location location, string value) : IHasLocation
+public readonly struct TableCell(Location location, string value) : IHasLocation
 {
     public Location Location { get; } = location;
     public string Value { get; } = value;