--- a/dotnet/Gherkin/Ast/TableRow.cs
+++ b/dotnet/Gherkin/Ast/TableRow.cs
@@ -1,6 +1,6 @@
 namespace Gherkin.Ast;
 
-public class TableRow(Location location, TableCell[] cells) : IHasLocation
+public class TableRow(Location location, IEnumerable<TableCell> cells) : IHasLocation
 {
     public Location Location { get; } = location;
     public IEnumerable<TableCell> Cells { get; } = cells;