… CHANGELOG.md
… dotnet/Gherkin/GherkinDialectProvider.cs
… dotnet/Gherkin/GherkinLine.cs
… dotnet/Gherkin/IGherkinLine.cs
… dotnet/Gherkin/ParserException.cs
… dotnet/Gherkin/TokenMatcher.cs
… dotnet/Gherkin/TokenScanner.cs