@obligaron 2025-01-14
> Maybe what we could do still is to take-over the [4 related unit tests](https://github.com/reqnroll/Reqnroll.VisualStudio/blob/main/Tests/Reqnroll.VisualStudio.Tests/Editor/Services/DeveroomGherkinParserTests.cs) from the VS extension to the Gherkin parser code. As it is now handled here, it is better to test it here. What do you think?

I think this is a good idea. I tried to add the tests but other implementations (c and perl) handle the errors differently then the other implementations (including .NET).